modify: JPA 조회 메서드 쿼리 중 일부 응답타입 수정

+++ src/main/java/com/waguwagu/weat/domain/analysis/model/entity/AnalysisResult.java
@@ -16,7 +16,7 @@ public class AnalysisResult {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "analysis_result_id")
-    private Integer analysisResultId;
+    private Long analysisResultId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "group_id", nullable = false)
